SPH-LBM second-order lifting and LBM-Stokes non-eq extrapolation
SPH-LBM coupling (CoupleSPHToLBM): upgraded from first-order
(equilibrium-only) to second-order lifting. The non-equilibrium part
of the LBM distribution is now reconstructed from the SPH stress
tensor: f_i = feq(rho, u) + fneq(stress). This preserves viscous
stress continuity across the interface, eliminating the artificial
boundary layer that first-order coupling created.

New SPH methods: InterpolateFields() for density/velocity at arbitrary
positions, InterpolateStressTensor() for the symmetric velocity
gradient tensor used by the lifting relation.

LBM-Stokes coupling (StokesToLBM): upgraded from constant-blend
equilibrium nudging to non-equilibrium extrapolation. Each LBM
boundary cell now gets feq from Stokes pressure/velocity plus
fneq copied from the current LBM state, preserving the viscous
stress structure while imposing correct macroscopic values.

Fixed all MSVC warnings (size_t to int, int to float conversions).

All 7 tests pass, zero warnings.

+++ include/mjwater/shallow_water.h
@@ -292,7 +292,7 @@ struct ShallowWaterSolver {
   // For y-sweep: primary = hv, transverse = hu, step along j with i fixed.
   template <typename IdxFn>
   void Sweep(float dt, uint32_t n_sweep, uint32_t n_perp,
-             IdxFn idx_fn, int ch_prim, int ch_trans) {
+             IdxFn idx_fn, size_t ch_prim, size_t ch_trans) {
     auto& h_data  = grid.channels[ch_h].data;
     auto& prim_data  = grid.channels[ch_prim].data;
     auto& trans_data = grid.channels[ch_trans].data;
